Logstash module.
