+++ cloudwatch-logs-to-logstash/.gitignore
@@ -2,3 +2,4 @@ lambda/.env
 lambda/deploy.env
 lambda/event.json
 lambda/config.json
+target
